move GcpApiManager to its own file
