+++ tools/run_tests/xds_k8s_test_driver/bin/run_td_setup.py
@@ -84,7 +84,7 @@ def main(argv):
     server_xds_host = xds_flags.SERVER_XDS_HOST.value
     server_xds_port = xds_flags.SERVER_XDS_PORT.value
 
-    gcp_api_manager = gcp.api.GcpApiManager()
+    gcp_api_manager = gcp.GcpApiManager()
 
     if security_mode is None:
         td = traffic_director.TrafficDirectorManager(gcp_api_manager,
