+++ tools/run_tests/xds_k8s_test_driver/framework/infrastructure/gcp/_internal/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2020 gRPC authors.
+# Copyright 2021 gRPC authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,7 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from framework.infrastructure.gcp import api
-from framework.infrastructure.gcp._internal import gcp_api_manager as _gcp_api_manager
-
-GcpApiManager = _gcp_api_manager.GcpApiManager
