+++ tools/run_tests/xds_k8s_test_driver/framework/infrastructure/gcp/gcp_flags.py
@@ -1,0 +1,31 @@
+# Copyright 2021 gRPC authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from absl import flags
+from googleapiclient import discovery
+
+PRIVATE_API_KEY_SECRET_NAME = flags.DEFINE_string(
+    "private_api_key_secret_name",
+    default=None,
+    help="Load Private API access key from the latest version of the secret "
+    "with the given name, in the format projects/*/secrets/*")
+V1_DISCOVERY_URI = flags.DEFINE_string("v1_discovery_uri",
+                                       default=discovery.V1_DISCOVERY_URI,
+                                       help="Override v1 Discovery URI")
+V2_DISCOVERY_URI = flags.DEFINE_string("v2_discovery_uri",
+                                       default=discovery.V2_DISCOVERY_URI,
+                                       help="Override v2 Discovery URI")
+COMPUTE_V1_DISCOVERY_FILE = flags.DEFINE_string(
+    "compute_v1_discovery_file",
+    default=None,
+    help="Load compute v1 from discovery file")
