+++ tools/run_tests/xds_k8s_test_driver/framework/infrastructure/gcp/network_security.py
@@ -44,7 +44,7 @@ class NetworkSecurityV1Alpha1(gcp.api.GcpStandardCloudApiResource):
         update_time: str
         create_time: str
 
-    def __init__(self, api_manager: gcp.api.GcpApiManager, project: str):
+    def __init__(self, api_manager: gcp.GcpApiManager, project: str):
         super().__init__(api_manager.networksecurity(self.api_version), project)
         # Shortcut to projects/*/locations/ endpoints
         self._api_locations = self.api.projects().locations()
